Don't strip the simulator system directory in newer SDKs
It's not necessary in Xcode 9, and the headers for the target and the
simulator actually differ subtly in a way that breaks compilation for
the simulator using the headers for the target device.

+++ strip-xcode.sh
@@ -45,5 +45,12 @@ cd iPhoneSimulator.platform
 rm -rf Developer/Library/CoreSimulator _CodeSignature
 cd Developer/SDKs/iPhoneSimulator.sdk
 rm -rf usr/share usr/lib usr/libexec Library Developer Applications
-rm -rf System
-cp -a ../../../../iPhoneOS.platform/Developer/SDKs/iPhoneOS.sdk/System .
+if [ ! -e System/Library/Frameworks/Foundation.framework/Foundation.tbd ]; then
+	# Xcode 7.x and 8.x has got full frameworks for the simulator here,
+	# replace them with thin frameworks with TBD files for the target.
+	# Xcode 9 has got TBD files for the simulator as well.
+	# On Xcode 9/iOS 11 SDK, building for the simulatorh with header
+	# files taken from the target breaks.
+	rm -rf System
+	cp -a ../../../../iPhoneOS.platform/Developer/SDKs/iPhoneOS.sdk/System .
+fi
